Вспомогательные функции перенесены в api.go из файлов с реализацией хендлера. Добавлены комментарии в файл api.go
